updated-code

+++ src/api/items/item_model.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-// use sqlx::types::BigDecimal;
+//  use sqlx::types::BigDecimal;
 use uuid::Uuid;
 
 #[derive(Serialize, Deserialize)]
@@ -8,7 +8,7 @@ use uuid::Uuid;
 pub struct ItemModel {
     pub item_id: Uuid,
     pub item_name: String,
-    pub item_price: i32,
+    pub item_price: f64,
     pub item_qty: i16,
     pub category_id: Uuid,
     pub created_at: Option<chrono::DateTime<chrono::Utc>>,
@@ -17,7 +17,7 @@ pub struct ItemModel {
 #[derive(Serialize, Deserialize)]
 pub struct CreateItemModel {
     pub item_name: String,
-    pub item_price: i32,
+    pub item_price: f64,
     pub item_qty: i16,
     pub category_id: Uuid,
 }
